fix(anlassraum): keep public input client bundle server-safe
